Moved the ConfigAccessor to the utils package.
* NEW: moved the ConfigAccessor to the utils package.
